requirements.txt
streamlit
python-dotenv
torch
transformers
whisper
langchain
langchain-community
langchain-chroma
pdfplumber
chromadb
requests